Ready

+++ all_lucky.py
@@ -7,11 +7,15 @@ from selenium.webdriver.support import expected_conditions as EC
 from webdriver_manager.chrome import ChromeDriverManager
 import time
 from datetime import datetime, timezone, timedelta
+import os  # Для проверки существования файла
 
 # Настройка Selenium
 options = webdriver.ChromeOptions()
+options.add_argument('--headless')
 options.add_argument('--no-sandbox')
 options.add_argument('--disable-dev-shm-usage')
+options.add_argument('--disable-gpu')
+options.add_argument('--window-size=1920,1080')
 
 url = "https://1wzjvm.top/casino/play/1play_1play_luckyjet"
 
@@ -90,10 +94,14 @@ def monitor_history(driver):
 
 # Основной код
 try:
-    # Открываем файл для записи заголовков, если файл пуст
-    with open("history_data.csv", mode="w", newline="") as file:
-        writer = csv.writer(file)
-        writer.writerow(["Type", "Value", "Time (MSK)"])  # Записываем заголовки
+    # Проверка, существует ли файл
+    file_exists = os.path.isfile("history_data.csv")
+
+    # Открытие файла для записи заголовков только если файл не существует
+    if not file_exists:
+        with open("history_data.csv", mode="w", newline="") as file:
+            writer = csv.writer(file)
+            writer.writerow(["Type", "Value", "Time (MSK)"])  # Записываем заголовки
 
     driver = initialize_browser()
     monitor_history(driver)
